Cp 90 scheduled tasks jira (#15)
* refactor(backend):added scheduled tasks for jira & sprint

* refactor(backend): added scheduled task for user stories

* refactor(backend): refactored sprint issues

* refactor(backend): removed creds

* refactor(backend):added scheduled tasks for jira & sprint

* refactor(backend): added scheduled task for user stories

* refactor(backend): refactored sprint issues

* refactor(backend): removed creds

* fix(data-model): fix bugs

* refactor(scheduled-task): fix scheduled task bug and refactor services

+++ cockpit_api/src/main/java/com/cockpit/api/model/dto/UserStoryDTO.java
@@ -14,7 +14,7 @@ public class UserStoryDTO {
 
     private Date doneDate;
 
-    private Integer storyPoint;
+    private double storyPoint;
 
     private String description;
 
@@ -64,11 +64,11 @@ public class UserStoryDTO {
         this.doneDate = doneDate;
     }
 
-    public Integer getStoryPoint() {
+    public double getStoryPoint() {
         return storyPoint;
     }
 
-    public void setStoryPoint(Integer storyPoint) {
+    public void setStoryPoint(double storyPoint) {
         this.storyPoint = storyPoint;
     }
